pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "omnis-py"
version = "0.1.0"
description = "Python library for Ex Libris Primo (OMNIS) API"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "TheUndefined" },
]
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "PyYAML>=6.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "respx>=0.21.0",
    "ruff>=0.3.0",
    "black>=24.0.0",
    "mypy>=1.9.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
omnis-cli = "omnis.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/omnis"]

[tool.black]
line-length = 120

[tool.ruff]
line-length = 120
target-version = "py311"
